Updating from origin/dev

+++ tests/Publish-DbaDacpac.Tests.ps1
@@ -6,9 +6,8 @@ Describe "$commandname Integration Tests" -Tags "IntegrationTests" {
 	BeforeAll {
 		$dbname = "dbatoolsci_publishdacpac"
 		$server = Connect-DbaInstance -SqlInstance $script:instance1
-		# Need a clean empty database
 		$null = $server.Query("Create Database [$dbname]")
-		$db = Get-DbaDatabase -SqlInstance $server -Database $dbname
+		$db = Get-DbaDatabase -SqlInstance $script:instance1 -Database $dbname
 		$null = $db.Query("CREATE TABLE dbo.example (id int); 
 			INSERT dbo.example
 			SELECT top 100 1 
@@ -22,11 +21,10 @@ Describe "$commandname Integration Tests" -Tags "IntegrationTests" {
 		Remove-DbaDatabase -SqlInstance $script:instance1, $script:instance2 -Database $dbname -Confirm:$false
 		Remove-Item -Confirm:$false -Path $publishprofile.FileName -ErrorAction SilentlyContinue
 	}
-	
-	Context "Testing pipability and that the command works" {
-		It -Skip "shows that the upate is complete" {
-			$results = $dacpac | Publish-DbaDacpac -PublishXml $publishprofile.FileName -Database $dbname -SqlInstance $script:instance2
-			$results.Result -match 'Update complete.' | Should Be $true
+	It "shows that the update is complete" {
+		$results = $dacpac | Publish-DbaDacpac -PublishXml $publishprofile.FileName -Database $dbname -SqlInstance $script:instance2
+		$results.Result -match 'Update complete.' | Should Be $true
+		if (($dacpac).Path) {
 			Remove-Item -Confirm:$false -Path ($dacpac).Path -ErrorAction SilentlyContinue
 		}
 	}
